refactor kafka event mappers structure
