+++ src/test/java/com/selina/lending/messaging/publisher/mapper/IPHelperTest.java
@@ -20,7 +20,7 @@ package com.selina.lending.messaging.publisher.mapper;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 
-import static com.selina.lending.messaging.publisher.mapper.IPHelper.getRemoteAddr;
+import static com.selina.lending.messaging.mapper.brokerrequest.IPHelper.getRemoteAddr;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class IPHelperTest {
